backend/requirements.txt
fastapi

django

uvicorn[standard]

pydantic-settings

python-multipart

python-dotenv

django-taggit

django-rest-framework

psycopg2-binary

djangorestframework-simplejwt